Caching the Inverse of a Matrix
Caching the Inverse of a Matrix. Matrix inversion is usually a costly
computation and there may be some benefit to caching the inverse of a
matrix rather than compute it repeatedly. Below are a pair of functions
that are used to create a special object that stores a matrix and caches
its inverse.
cachesolve  function computes the inverse of the special "matrix"
created by makeCacheMatrix above. If the inverse has already been
calculated (and the matrix has not changed), then it should retrieve the
inverse from the cache.